pdf-service/requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
pytesseract
Pillow
python-multipart
